backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-pptx==1.0.2
python-multipart==0.0.20
